requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dill==0.3.9
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lockfile==0.12.2
luigi==3.6.0
pyfunctional==1.5.0
pystac==1.11.0
pystac-client==0.8.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
s3cmd==2.4.0
six==1.17.0
tabulate==0.9.0
tenacity==8.5.0
tornado==6.4.2
typing_extensions==4.12.2
urllib3==2.3.0
